sqlite2pg/requirements.txt
flake8==4.0.1
psycopg2-binary==2.9.3
python_log_indenter==0.9
emoji==1.7.0
python-dateutil==2.8.2